naming and structure refactor
