+++ cron-saturas/data/sampleTransformedPlotPayload.json
@@ -1,0 +1,20 @@
+[
+    {
+        "date": "6/9/2020 12:00:00 AM",
+        "swpAvg": "22.995",
+        "et": "4.0",
+        "solarRadiation": "5283.0",
+        "dailyMaxWind": "5.2",
+        "rain": "0.3",
+        "irrigationValue": "1.19"
+    },
+    {
+        "date": "6/9/2020 12:00:00 AM",
+        "swpAvg": "22.995",
+        "et": "4.0",
+        "solarRadiation": "5283.0",
+        "dailyMaxWind": "5.2",
+        "rain": "0.3",
+        "irrigationValue": "1.19"
+    }
+]
