+++ cron-saturas/data/sampleTransformedSensorPayload.json
@@ -1,0 +1,10 @@
+[
+    {
+        "date": "6/9/2020 12:00:00 AM",
+        "sensorSwpAvg": "8.74"
+    },
+    {
+        "date": "6/9/2020 12:00:00 AM",
+        "sensorSwpAvg": " 8.74"
+    }
+]
