+++ cron-saturas/data/sampleTransformedTransmitorPayload.json
@@ -1,0 +1,18 @@
+[
+    {
+        "date": "6/9/2020 12:00:00 AM",
+        "volt": "3.637",
+        "rssi": "-111.0",
+        "txTemp": "14.4",
+        "txRh": "44.0",
+        "sf": "10.0"
+    },
+    {
+        "date": "6/9/2020 12:00:00 AM",
+        "volt": "3.637",
+        "rssi": "-111.0",
+        "txTemp": "14.4",
+        "txRh": "44.0",
+        "sf": "10.0"
+    }
+]
