+++ helpers/saturasObjectModels.js
@@ -1,0 +1,19 @@
+module.exports = {
+
+    dataObject : {
+        canal : {
+            name: 'canal',
+            sensorId: 18165,
+            plotTIndex: 1332,
+            transmitorId: 'A81758FFFE04C5C4'
+        },
+    
+        poundstone : {
+            name: 'poundstone',
+            sensorId: 15564,
+            plotTIndex: 1368,
+            transmitorId: 'A81758FFFE04C55D'
+        }
+    }
+}  
+
