--- jainLogicData/debugging/README.md
@@ -1,1 +1,0 @@
-DO NOT REMOVE.  Place holder to include this directory in source control.
