+++ package.json
@@ -5,10 +5,10 @@
   "main": "app.js",
   "scripts": {
     "start": "nodemon -r dotenv/config app.js",
-    "getJain:local": "node -r dotenv/config getJainLogicData.js",
-    "getJain:prod": "node getJainLogicData.js",
-    "getSat:local": "node -r dotenv/config getSaturasData.js",
-    "getSat:prod": "node getSaturasData.js",
+    "cron:jainlogic:local": "node -r dotenv/config cron-jainlogic.js",
+    "cron:jainlogic:prod": "node cron-jainlogic.js",
+    "cron:saturas:local": "node -r dotenv/config cron-saturas.js",
+    "cron:saturas:prod": "node cron-saturas.js",
     "devtest:local": "node -r dotenv/config dev.js"
   },
   "author": "Matthew McGuff",
