--- saturasData/data/sampleTransformedSensorPayload.json
@@ -1,10 +1,0 @@
-[
-    {
-        "date": "6/9/2020 12:00:00 AM",
-        "sensorSwpAvg": "8.74"
-    },
-    {
-        "date": "6/9/2020 12:00:00 AM",
-        "sensorSwpAvg": " 8.74"
-    }
-]
